Passport-Recogintion-Main/Passport-Recognition-Server/requirements.txt
flask==2.0.2
flask_restful==0.3.9
flask_sqlalchemy==2.5.1
flask_cors==3.0.10
PyJWT==2.1.0
PassportEye~=2.1.0
Werkzeug~=2.0.2
pymysql~=1.0.2
pytesseract~=0.3.8
opencv-python~=4.5.4.60
Pillow~=8.4.0